backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
aiohttp
sentence-transformers
numpy
google-cloud-aiplatform
vertexai
